requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
beautifulsoup4==4.12.2
selenium==4.15.2
openai==1.3.7
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
redis==5.0.1
celery==5.3.4
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
requests==2.31.0
lxml==4.9.3
fake-useragent==1.4.0
asyncio-throttle==1.0.2
playwright==1.40.0
aiohttp==3.9.1 